+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.10:
   * Auto search in all layer (except for Chinese) feature is added when no match is found
   * Default Keyboard size increased
+  * Mongo performance optimizations using hash matching instead of the whole string
 
 ## 6.9:
   * Added function to hide unmatching layers when searching across all layers