+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Auto search in all layer (except for Chinese) feature is added when no match is found
   * Default Keyboard size increased
   * Mongo performance optimizations using hash matching instead of the whole string
+  * Cohorts updated to only show when not empty
 
 ## 6.9:
   * Added function to hide unmatching layers when searching across all layers